Move tool search RPC to dedicated processor
