+++ codex-rs/app-server/src/request_processors.rs
@@ -517,6 +517,7 @@ mod remote_control_processor;
 mod search;
 mod thread_processor;
 mod token_usage_replay;
+mod tool_search_processor;
 mod turn_processor;
 mod windows_sandbox_processor;
 
@@ -540,6 +541,7 @@ pub(crate) use remote_control_processor::RemoteControlRequestProcessor;
 pub(crate) use search::SearchRequestProcessor;
 pub(crate) use thread_goal_processor::ThreadGoalRequestProcessor;
 pub(crate) use thread_processor::ThreadRequestProcessor;
+pub(crate) use tool_search_processor::ToolSearchRequestProcessor;
 pub(crate) use turn_processor::TurnRequestProcessor;
 pub(crate) use windows_sandbox_processor::WindowsSandboxRequestProcessor;
